+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+### v0.13.7
+- Allow input types to be passed in from variables using \stdClass instead of associative arrays (#535)
+
 #### v0.13.6
 - QueryPlan can now be used on interfaces not only objects. (#495)
 - Array in variables in place of object shouldn't cause fatal error (fixes #467)